all_DAGs: deal with disconnected graphs
The empty edge set was discarded, now we include it. In general, the
handling of disconnected graphs was broken: if the vertices k+1, ..., n
are all isolated in a given edge set, then the graph was created on only
k vertices instead which results in incorrect CI structures later on.